+++ packages/tools/dom/rollup.config.js
@@ -13,20 +13,22 @@ import {
 
 import pkg from "./package.json";
 
-const input = pkg.module;
-const banner = makeBanner(pkg);
 const analyzer = analyze({
-  limit: 20,
+  limit: 15,
   root: path.resolve('../../../'),
   stdout: true,
   summaryOnly: true
 });
+const banner = makeBanner(pkg);
+const external = pkg.peerDependencies && Object.keys(pkg.peerDependencies) || [];
+const input = pkg.module;
 const treeshake = {
   annotations: true,
   moduleSideEffects: false,
 };
 
 const cjsConfig = {
+    external,
     input,
     output: {
         banner,
@@ -37,13 +39,13 @@ const cjsConfig = {
     plugins: [
         resolve(),
         commonjs(),
-        cleanup(),
-        analyzer
+        cleanup()
     ],
     treeshake
 };
 
 const browserConfig = {
+    external,
     input,
     output: {
         banner,
@@ -56,8 +58,7 @@ const browserConfig = {
         resolve(),
         commonjs(),
         cleanup(),
-        buble(),
-        analyzer
+        buble()
     ],
     treeshake
 };
@@ -74,7 +75,8 @@ const browserMinifiedConfig = {
             output: {
                 preamble: banner
             }
-        })
+        }),
+        analyzer
     ],
     treeshake
 };