+++ packages/tools/geo/rollup.config.js
@@ -13,23 +13,26 @@ import {
 
 import pkg from "./package.json";
 
-const input = pkg.module;
-const banner = makeBanner(pkg);
 const analyzer = analyze({
-  limit: 10,
+  limit: 15,
   root: path.resolve('../../../'),
   stdout: true,
   summaryOnly: true
 });
+const banner = makeBanner(pkg);
+const external = pkg.peerDependencies && Object.keys(pkg.peerDependencies) || [];
+const input = pkg.module;
 const treeshake = {
   annotations: true,
-  moduleSideEffects: id => !(
-    /@svizzle\/file/g.test(id) ||
-    /lamb/g.test(id)
-  ),
+  moduleSideEffects: [
+    // bin/featuresDirToCollection.js, unused by the lib
+    '@svizzle/file',
+    'commander'
+  ],
 };
 
 const cjsConfig = {
+    external,
     input,
     output: {
         banner,
@@ -41,12 +44,12 @@ const cjsConfig = {
         resolve(),
         commonjs(),
         cleanup(),
-        analyzer
     ],
     treeshake
 };
 
 const browserConfig = {
+    external,
     input,
     output: {
         banner,
@@ -60,7 +63,6 @@ const browserConfig = {
         commonjs(),
         cleanup(),
         buble(),
-        analyzer
     ],
     treeshake
 };
@@ -77,7 +79,8 @@ const browserMinifiedConfig = {
             output: {
                 preamble: banner
             }
-        })
+        }),
+        analyzer
     ],
     treeshake
 };