+++ packages/tools/geometry/rollup.config.js
@@ -13,20 +13,22 @@ import {
 
 import pkg from "./package.json";
 
-const input = pkg.module;
-const banner = makeBanner(pkg);
 const analyzer = analyze({
-  limit: 10,
+  limit: 15,
   root: path.resolve('../../../'),
   stdout: true,
   summaryOnly: true
 });
+const banner = makeBanner(pkg);
+const external = pkg.peerDependencies && Object.keys(pkg.peerDependencies) || [];
+const input = pkg.module;
 const treeshake = {
   annotations: true,
   moduleSideEffects: false,
 };
 
 const cjsConfig = {
+    external,
     input,
     output: {
         banner,
@@ -38,12 +40,12 @@ const cjsConfig = {
         resolve(),
         commonjs(),
         cleanup(),
-        analyzer
     ],
     treeshake
 };
 
 const browserConfig = {
+    external,
     input,
     output: {
         banner,
@@ -57,7 +59,6 @@ const browserConfig = {
         commonjs(),
         cleanup(),
         buble(),
-        analyzer
     ],
     treeshake
 };
@@ -74,7 +75,8 @@ const browserMinifiedConfig = {
             output: {
                 preamble: banner
             }
-        })
+        }),
+        analyzer
     ],
     treeshake
 };